CreateTests_ImportLayout_And_ParkingPrice

+++ ParkNetApp/Data/Entities/NonSubscriptionParkingTariff.cs
@@ -3,8 +3,8 @@
 public class NonSubscriptionParkingTariff  // Entidade isolada de tarifas de estacionamento não subscritas (sem avença)
 {
     public int Id { get; set; }
-    public int Limit { get; set; }
-    public decimal Tariff { get; set; }
+    public double Limit { get; set; }
+    public double Tariff { get; set; }
     public DateOnly ActiveSince { get; set; }
     public DateOnly? ActiveUntil { get; set; }
 }
